Very WIP aurora backend + scheduler refactor
