+++ tests/test_aurora.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 import jinja_dicts
-import herc.aurora as aurora
+import herc.aurorasched as scheduler
 import json
 
 class TestAurora(TestCase):
@@ -9,7 +9,7 @@ class TestAurora(TestCase):
 		with open('tests/data/full_submit.json', 'r') as fullsub:
 			fullrq = json.load(fullsub)
 
-		jinjadict = aurora.build_jinja_dict("TESTJOB", fullrq)
+		jinjadict = scheduler.build_jinja_dict("TESTJOB", fullrq)
 		self.assertEqual( jinjadict, jinja_dicts.full_submit )
 
 	def build_mock_jobstatus(self, statuses, failmsg = "PLACEHOLDER: Job failed."):
@@ -22,22 +22,22 @@ class TestAurora(TestCase):
 	def test_nonterm_status(self):
 		"""Test that we correctly report nonterminal statuses."""
 		nonterm_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "RUNNING" ])
-		self.assertEqual( aurora.determine_true_status(nonterm_status)[0], "RUNNING" )
+		self.assertEqual( scheduler.determine_true_status(nonterm_status)[0], "RUNNING" )
 
 	def test_lost_status(self):
 		"""Test that we report lost jobs as rescheduled if they didn't get killed."""
 		lost_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "LOST" ])
-		self.assertEqual( aurora.determine_true_status(lost_status)[0], "RESCHEDULED" )
+		self.assertEqual( scheduler.determine_true_status(lost_status)[0], "RESCHEDULED" )
 
 	def test_killed_status(self):
 		"""Test that we report user killed jobs as killed."""
 		killed_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "KILLING", "KILLED" ])
-		self.assertEqual( aurora.determine_true_status(killed_status)[0], "KILLED" )
+		self.assertEqual( scheduler.determine_true_status(killed_status)[0], "KILLED" )
 
 	def test_lost_killed_status(self):
 		"""Test that we report user killed jobs as killed, even if they get lost."""
 		lost_killed_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "KILLING", "LOST" ])
-		self.assertEqual( aurora.determine_true_status(lost_killed_status)[0], "KILLED" )
+		self.assertEqual( scheduler.determine_true_status(lost_killed_status)[0], "KILLED" )
 
 	def test_kill_failed_status(self):
 		"""Test that we report user killed jobs as killed, even if they complete before the kill request gets through."""
@@ -45,28 +45,28 @@ class TestAurora(TestCase):
 		#finishes or fails before the kill request arrives. In this case, it proceeds to finished or failed.
 		#For simplicity, we tell users that the job was successfully killed, even if it did something else.
 		kill_failed_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "KILLING", "FAILED" ])
-		self.assertEqual( aurora.determine_true_status(kill_failed_status)[0], "KILLED" )
+		self.assertEqual( scheduler.determine_true_status(kill_failed_status)[0], "KILLED" )
 
 	def test_preempting_status(self):
 		"""Test that we report PREEMPTING -> KILLED jobs as rescheduled."""
 		preempting_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "PREEMPTING", "KILLED" ])
-		self.assertEqual( aurora.determine_true_status(preempting_status)[0], "RESCHEDULED" )
+		self.assertEqual( scheduler.determine_true_status(preempting_status)[0], "RESCHEDULED" )
 
 	def test_restarting_status(self):
 		"""Test that we report RESTARTING -> KILLED jobs as rescheduled."""
 		restarting_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "RESTARTING", "KILLED" ])
-		self.assertEqual( aurora.determine_true_status(restarting_status)[0], "RESCHEDULED" )
+		self.assertEqual( scheduler.determine_true_status(restarting_status)[0], "RESCHEDULED" )
 
 	def test_failed_status(self):
 		"""Test that we report terminal failed jobs as such."""
 		failed_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "RUNNING", "FAILED" ])
-		self.assertEqual( aurora.determine_true_status(failed_status)[0], "FAILED" )
+		self.assertEqual( scheduler.determine_true_status(failed_status)[0], "FAILED" )
 
 	def test_mem_exceeded(self):
 		"""Test that we report memory exceeded jobs as such, and correctly pull out requested and used limits"""
 		failed_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "RUNNING", "FAILED" ],
 		                                          failmsg = "Memory limit exceeded: Requested 128MB, Used 130MB." )
-		aurora_status = aurora.determine_true_status(failed_status)
+		aurora_status = scheduler.determine_true_status(failed_status)
 		self.assertEqual( aurora_status[0], "MEM_EXCEEDED" )
 		self.assertEqual( aurora_status[1]['requested'], "128MB" )
 		self.assertEqual( aurora_status[1]['used'], "130MB" )
@@ -75,7 +75,7 @@ class TestAurora(TestCase):
 		"""Test that we report disk exceeded jobs as such, and correctly pull out requested and used limits"""
 		failed_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "RUNNING", "FAILED" ],
 		                                          failmsg = "Disk limit exceeded.  Reserved 1234 bytes vs used 2345 bytes." )
-		aurora_status = aurora.determine_true_status(failed_status)
+		aurora_status = scheduler.determine_true_status(failed_status)
 		self.assertEqual( aurora_status[0], "DISK_EXCEEDED" )
 		self.assertEqual( aurora_status[1]['requested'], "1234BYTES" )
 		self.assertEqual( aurora_status[1]['used'], "2345BYTES" )
@@ -85,4 +85,4 @@ class TestAurora(TestCase):
 		failed_status = self.build_mock_jobstatus([ "INIT", "PENDING", "ASSIGNED", "RUNNING", "FAILED" ],
 		                                          failmsg = "Memory limit exceeded: asked for 128MB, spent 130MB." )
 		with self.assertRaises(AssertionError):
-			aurora.determine_true_status(failed_status)
+			scheduler.determine_true_status(failed_status)
